# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Opsgate API",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/activity": {
      "get": {
        "tags": [
          "observability"
        ],
        "operationId": "get_activity",
        "parameters": [
          {
            "name": "channel",
            "in": "query",
            "description": "Filter by channel: api, mcp, browser, system",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tool",
            "in": "query",
            "description": "Alias for action filter",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "outcome",
            "in": "query",
            "description": "Filter by outcome: ok, denied, error",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "credential",
            "in": "query",
            "description": "Filter by target credential alias",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size 1-100",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Continuation cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recent activity timeline",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/audit/events": {
      "get": {
        "tags": [
          "observability"
        ],
        "operationId": "get_audit_events",
        "parameters": [
          {
            "name": "channel",
            "in": "query",
            "description": "Filter by channel",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "in": "query",
            "description": "Filter by audit action",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "outcome",
            "in": "query",
            "description": "Filter by outcome",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "target_type",
            "in": "query",
            "description": "Filter by target type",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "credential",
            "in": "query",
            "description": "Filter by target credential alias",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size 1-100",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Continuation cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed audit events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/history/api-calls": {
      "get": {
        "tags": [
          "observability"
        ],
        "operationId": "get_api_call_history",
        "parameters": [
          {
            "name": "channel",
            "in": "query",
            "description": "Filter by channel",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "credential",
            "in": "query",
            "description": "Filter by credential alias",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "outcome",
            "in": "query",
            "description": "Filter by outcome",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "error_kind",
            "in": "query",
            "description": "Filter by safe error kind",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size 1-100",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Continuation cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API call history rows",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiCallHistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/api/v1/summary": {
      "get": {
        "tags": [
          "observability"
        ],
        "operationId": "get_summary",
        "responses": {
          "200": {
            "description": "Credential and activity summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid Bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ActivityItem": {
        "type": "object",
        "required": [
          "id",
          "created_at",
          "action",
          "channel",
          "outcome",
          "severity"
        ],
        "properties": {
          "action": {
            "type": "string"
          },
          "channel": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "outcome": {
            "type": "string"
          },
          "purpose": {
            "type": [
              "string",
              "null"
            ]
          },
          "severity": {
            "type": "string"
          },
          "target_key": {
            "type": [
              "string",
              "null"
            ]
          },
          "target_type": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ActivityResponse": {
        "type": "object",
        "required": [
          "items",
          "page"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityItem"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageResponse"
          }
        }
      },
...
      "ApiCallHistoryResponse": {
        "type": "object",
        "required": [
          "calls",
          "page"
        ],
        "properties": {
          "calls": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiCallHistoryItem"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageResponse"
          }
        }
      },
...
      "AuditEventsResponse": {
        "type": "object",
        "required": [
          "events",
          "page"
        ],
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditEventResponse"
            }
          },
          "page": {
            "$ref": "#/components/schemas/PageResponse"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "hint": {
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "SummaryResponse": {
        "type": "object",
        "required": [
          "credentials",
          "activity_24h"
        ],
        "properties": {
          "activity_24h": {
            "$ref": "#/components/schemas/ActivitySummaryResponse"
          },
          "credentials": {
            "$ref": "#/components/schemas/CredentialSummaryResponse"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}